made noise level consistent with explained variance in PCA decomposition
Noise calculation is now based of explained variance parameter for the PCA decomposition, which is simpler.
Also the sampler is now checking the quality of the prediction without taking noise into account since adding points will not help if the error is already noise dominated. Instead it currently warns that the corresponding precision parameters should be reduced.
In addition to explained_variance_cutoff there is a new parameter noise_percentage which can be reduced without affecting the PCA.