fallback to normal GPs if sparse does not converge
Implemented a fallback to normal GPs if the sparse GP does not converge with reduced points. This uses the data slightly more efficient and the normal GPs are easier to train leading to better hyperparameters and a better fit to the data.

Also updated the Readme and fixed a few minor inconsistencies in the input files