new error estimator to be tested
I have added a new error estimator based on a less random sampling of the support points. This estimator can be exakt given N_points > number of GP components and it estimates the variance faithfully assuming that the local likeleehood is well-behaved. It should get rid of most random outliers of the previous method.

I also replaced the white_noise_level by a white_noise_ratio allowing the user to specify the withe noise error in realtion to the PCA error. A value of one reproduces the old behaviour, while a value of 0.5 or 0.1 makes the white noise term less relevant without fully removing it.